Fixed memory leaks and truncation warnings. Added slider for spaceship mass.

+++ Engine.h
@@ -19,7 +19,7 @@ private:
 	Scene m_scene;
 	ImGuiManager m_imguiManager;
 	bool m_Running;
-	float m_time;
+	long double m_time;
 	int fps;
 private:
 	void Update();
